+++ molo/surveys/tests/constants.py
@@ -44,7 +44,7 @@ SEGMENT_FORM_DATA = {
     'surveys_surveyresponserule_related-INITIAL_FORMS': ['0'],
     'surveys_surveyresponserule_related-MIN_NUM_FORMS': ['0'],
     'surveys_surveyresponserule_related-MAX_NUM_FORMS': ['1000'],
-    'surveys_surveyresponserule_related-0-survey': ['10'],
+    'surveys_surveyresponserule_related-0-survey': ['11'],
     'surveys_surveyresponserule_related-0-id': [''],
     'surveys_surveyresponserule_related-0-DELETE': [''],
     'surveys_groupmembershiprule_related-TOTAL_FORMS': ['0'],